pyproject.toml
[tool.poetry]
name = "ami_meeting_svc"
version = "1.0.0"
description = ""
authors = ["ThesenAI"]
readme = "README.md"
packages = [{include = "ami_meeting_svc", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.2"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.28.1"

[tool.poetry.scripts]
ami_meeting_svc = "ami_meeting_svc.main:main"

[tool.pytest.ini_options]
pythonpath = [ "src/" ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
